refactor: Message Enum 추가, 예외 처리 코드 수정

+++ src/main/java/nextstep/schedule/ScheduleService.java
@@ -1,15 +1,19 @@
 package nextstep.schedule;
 
+import com.sun.jdi.request.DuplicateRequestException;
 import nextstep.reservation.Reservation;
 import nextstep.reservation.ReservationDao;
 import nextstep.theme.Theme;
 import nextstep.theme.ThemeDao;
 import org.springframework.stereotype.Service;
 
-import javax.naming.ContextNotEmptyException;
+import javax.management.openmbean.KeyAlreadyExistsException;
+import javax.naming.NotContextException;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
+
+import static nextstep.config.Messages.*;
 
 @Service
 public class ScheduleService {
@@ -25,8 +29,11 @@ public class ScheduleService {
 
     public Long create(ScheduleRequest scheduleRequest) {
         Optional<List<Theme>> themeList = themeDao.findById(scheduleRequest.getThemeId());
-        if (themeList.get().isEmpty()) {
-            throw new NullPointerException("Cannot Find Theme");
+        if (themeList.isEmpty() || themeList.get().isEmpty()) {
+            throw new NullPointerException(THEME_NOT_FOUND.getMessage());
+        }
+        if (scheduleDao.isExistsByTimeAndDate(scheduleRequest.getTime(), scheduleRequest.getDate())) {
+            throw new DuplicateRequestException(ALREADY_REGISTERED_SCHEDULE.getMessage());
         }
         return scheduleDao.save(scheduleRequest.toEntity(themeList.get().get(0)));
     }
@@ -35,10 +42,13 @@ public class ScheduleService {
         return scheduleDao.findByThemeIdAndDate(themeId, date);
     }
 
-    public void deleteById(Long id) throws ContextNotEmptyException {
+    public void deleteById(Long id) throws NotContextException {
         List<Reservation> reservationList = reservationDao.findByScheduleId(id);
+        if (scheduleDao.findById(id).isEmpty()){
+            throw new NotContextException(SCHEDULE_NOT_FOUND.getMessage());
+        }
         if (Objects.requireNonNull(reservationList).size() != 0) {
-            throw new ContextNotEmptyException("테마에 예약이 존재함");
+            throw new KeyAlreadyExistsException(ALREADY_REGISTERED_RESERVATION.getMessage());
         }
         scheduleDao.deleteById(id);
     }
