+++ src/main/java/nextstep/theme/ThemeService.java
@@ -1,5 +1,6 @@
 package nextstep.theme;
 
+import com.sun.jdi.request.DuplicateRequestException;
 import nextstep.schedule.ScheduleDao;
 import org.springframework.stereotype.Service;
 
@@ -7,6 +8,8 @@ import javax.management.openmbean.KeyAlreadyExistsException;
 import javax.naming.NotContextException;
 import java.util.List;
 import java.util.Optional;
+
+import static nextstep.config.Messages.*;
 
 @Service
 public class ThemeService {
@@ -20,7 +23,7 @@ public class ThemeService {
 
     public Long create(ThemeRequest themeRequest) {
         if (themeDao.findByNameAndPrice(themeRequest)) {
-            throw new KeyAlreadyExistsException("Already Registered Theme");
+            throw new DuplicateRequestException(ALREADY_REGISTERED_THEME.getMessage());
         }
         return themeDao.save(themeRequest.toEntity());
     }
@@ -31,11 +34,11 @@ public class ThemeService {
 
     public void delete(Long id) throws NotContextException {
         Optional<List<Theme>> themeList = themeDao.findById(id);
-        if (themeList.get().isEmpty()) {
-            throw new NotContextException("Theme not registered");
+        if (themeList.isEmpty() || themeList.get().isEmpty()) {
+            throw new NotContextException(THEME_NOT_FOUND.getMessage());
         }
         if (scheduleDao.isExistsByThemeId(id)) {
-            throw new KeyAlreadyExistsException("Exists Registered scheduled");
+            throw new KeyAlreadyExistsException(ALREADY_REGISTERED_SCHEDULE.getMessage());
         }
         themeDao.delete(id);
     }
